25/BranchCleanupAndFormImplementation
Removed unnecesary files from branch

Implemented AccountProfileDiverseModel

Implemented AccountProfileModel

Implemented AuthUserModel

Implemented ChangePasswordModel

Implemented CompanyModel

Implemented JobAdvertModel

Implemented JobAdvertPaginationModel

Implemented RegisterUserModel

Implemented VacantJobModel
