+++ JobAgentSolutions/BlazorWebsite/Data/FormModels/AccountProfileDiverseModel.cs
@@ -1,0 +1,8 @@
+﻿namespace BlazorWebsite.Data.FormModels
+{
+    public class AccountProfileDiverseModel
+    {
+        public int ConsultantAreaIdToBeAssigned { get; set; }
+        public int ConsultantAreaIdToBeRemoved { get; set; }
+    }
+}
