+++ JobAgentSolutions/BlazorWebsite/Data/FormModels/AuthUserModel.cs
@@ -1,0 +1,32 @@
+﻿using System.ComponentModel.DataAnnotations;
+using System.Net.Mail;
+
+namespace BlazorWebsite.Data.FormModels
+{
+    public class AuthUserModel
+    {
+        [Required(AllowEmptyStrings = false, ErrorMessage = "Email adresse er påkrævet.")]
+        [StringLength(255, ErrorMessage = "Email er for long (255 karakter begrænse).")]
+        [EmailAddress(ErrorMessage = "Indtast en gyldig email adresse.")]
+        [DataType(DataType.EmailAddress)]
+        public string Email { get; set; }
+
+        [Required(AllowEmptyStrings = false, ErrorMessage = "Adgangskode er påkrævet.")]
+        [StringLength(255)]
+        [DataType(DataType.Password)]
+        public string Password { get; set; }
+
+        public bool IsValidEmail(string email)
+        {
+            try
+            {
+                var addr = new MailAddress(email);
+                return addr.Address == email;
+            }
+            catch
+            {
+                return false;
+            }
+        }
+    }
+}
