--- JobAgentSolutions/BlazorWebsite/Data/Providers/IRefreshProvider.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Threading.Tasks;
-
-namespace BlazorWebsite.Data.Providers
-{
-    public interface IRefreshProvider
-    {
-        event Func<Task> RefreshRequest;
-
-        void CallRefreshRequest();
-    }
-}
