--- JobAgentSolutions/BlazorWebsite/Pages/Administrate/CompanyPage.razor
@@ -1,84 +1,0 @@
-﻿@page "/admin/settings/companies"
-@attribute [Authorize(Roles = "Konsulent, System Administrator, Udvikler")]
-@using BlazorWebsite.Shared.Components.Diverse;
-@using BlazorWebsite.Shared.Components.Modals.CompanyModals;
-
-
-<PageDocumentTitle DocumentTitle="Administrer - Virksomheds Indstillinger" />
-
-<div class="d-flex">
-    <PageHeadingTitle DisplayTitle="Virksomheds Indstillinger" />
-
-    <div class="d-inline-block px-4 pt-4">
-        <a class="btn btn-sm btn-info cursor-pointer text-white" data-bs-toggle="modal" data-bs-target="#ModalCreateCompany">
-            Opret ny Virksomhed
-        </a>
-    </div>
-</div>
-
-
-@if (_companies != null)
-{
-    @if (_isLoadingData)
-    {
-        <MessageAlert IsLoading="_isLoadingData" Message="Indlæser data..."></MessageAlert>
-    }
-    else
-    {
-        @if (_companies.Count() == 0)
-        {
-            <MessageAlert Alert="MessageAlert.AlertType.Warning"
-                          Message="Der blev ikke fundet nogen virksomheder."
-                          MessageOptional="Opret virksomheder, ved at trykke på knappen : Opret ny Virksomhed." />
-        }
-        else
-        {
-            <div class="table-responsive table-responsive-sm table-responsive-md">
-                <table class="table table-striped table-hover">
-                    <caption>Liste over virksomheder</caption>
-                    <thead class="table-primary text-light">
-                        <tr>
-                            <th scope="col">#</th>
-                            <th scope="col">Navn</th>
-                            <th scope="col"></th>
-                            <th scope="col"></th>
-                        </tr>
-                    </thead>
-                    <tbody>
-
-                        @foreach (var company in _companies)
-                        {
-                            <tr id="@(company.Id)">
-                                <th scope="row">@(company.Id)</th>
-                                <td>@(company.Name)</td>
-                                <td>
-                                    <span @onclick="(OnClickEvent => OnClick_OpenEditModal(company.Id))" data-bs-toggle="modal" data-bs-target="#ModalEditCompany">
-                                        <Popover Title="Ændre information" Body="Ved denne funktion, kan du ændre informationen på denne virksomehd." Color="Popover.PopoverType.Info" />
-                                    </span>
-                                </td>
-                                <td>
-                                    <span @onclick="(OnClickEvent => OnClick_RemoveCompanyModal(company.Id))" data-bs-toggle="modal" data-bs-target="#ModalRemoveCompany">
-                                        <Popover Title="Forsigtig!" Body="Denne knap sletter det valgte opslag, forevigt!" Color="Popover.PopoverType.Danger" />
-                                    </span>
-                                </td>
-                            </tr>
-                        }
-
-                    </tbody>
-                </table>
-            </div>
-        }
-
-        <CreateCompanyModal />
-
-        <EditCompanyModal CompanyModel="_companyModel" />
-
-        <RemoveCompanyModal CompanyId="_companyId" />
-    }
-}
-else
-{
-    <MessageAlert Alert="MessageAlert.AlertType.Error"
-                  Message="Prøv at genindlæs siden."
-                  MessageOptional="@(errorMessage)" />
-}
