--- JobAgentSolutions/BlazorWebsite/Pages/Dashboard/Account/ChangePassword.razor
@@ -1,61 +1,0 @@
-﻿@page "/admin/account/change-password"
-@*@attribute [Authorize(Roles = "Konsulent, System Administrator, Udvikler")]*@
-@using BlazorWebsite.Shared.Components.Diverse;
-
-<PageDocumentTitle DocumentTitle="Bruger Indstillinger - Ændre Adgangskode" />
-
-<div class="page-content p-4">
-
-    <PageHeadingTitle DisplayTitle="Skift Kodeord" />
-
-    @if (_isLoadingData)
-    {
-        <MessageAlert Message="Indlæser nødvendig data; vent venligst." Alert="MessageAlert.AlertType.Info" IsLoading="_isLoadingData" />
-    }
-
-
-    @if (!string.IsNullOrEmpty(_infoMessage))
-    {
-        <MessageAlert FullWidth="false" Alert="MessageAlert.AlertType.Success" Message="@_infoMessage" />
-    }
-
-    @if (!string.IsNullOrEmpty(_errorMessage))
-    {
-        <MessageAlert FullWidth="false" Alert="MessageAlert.AlertType.Error" Message="@_errorMessage" />
-    }
-
-
-    <div class="form-group">
-        <label for="fakeInputEmail" class="admin-label">Email Adresse :</label>
-        <span id="fakeInputEmail" type="text" class="form-control form-control-sm" readonly>@(changePasswordModel.Email)</span>
-    </div>
-
-    <EditForm Model="@changePasswordModel" OnValidSubmit="OnValidSubmit_ChangeUserPassword" OnInvalidSubmit="OnInvalidSubmit_ChangeUserPassword">
-
-        <DataAnnotationsValidator />
-
-        <div class="form-group">
-            <label for="inputPassword" class="admin-label">Ny Adgangskode :</label>
-            <input id="inputPassword" type="password" class="form-control form-control-sm" placeholder="Indtast adgangskode.." @bind="changePasswordModel.Password" />
-            <ValidationMessage For="@(() => changePasswordModel.Password)" />
-        </div>
-        <div class="form-group">
-            <label for="inputConfirmPassword" class="admin-label">Bekræft Adgangskode :</label>
-            <input id="inputConfirmPassword" type="password" class="form-control form-control-sm" placeholder="Bekræft adgangskode.." @bind="changePasswordModel.ConfirmPassword" />
-            <ValidationMessage For="@(() => changePasswordModel.ConfirmPassword)" />
-            <ValidationSummary Model="@changePasswordModel" />
-        </div>
-        <button type="submit" class="btn btn-outline-primary mt-2" disabled="@isProcessingPasswordChangeRequest">
-            @if (isProcessingPasswordChangeRequest)
-            {
-                <span class="spinner-border spinner-border-sm" role="status" aria-hidden="true"></span>
-                <span>Skifter adgangskode...</span>
-            }
-            else
-            {
-                <span>Skift adgangskode</span>
-            }
-        </button>
-
-    </EditForm>
-</div>
