--- JobAgentSolutions/BlazorWebsite/Pages/JobCardPage.razor
@@ -1,47 +1,0 @@
-﻿@page "/job/uncategorized"
-@page "/job/{CategoryId:int}"
-@page "/job/{CategoryId:int}/{SpecializationId:int}"
-
-<div class="container-fluid my-3">
-    @if (_isLoadingData)
-    {
-        <MessageAlert Message="Indlæser Stillingsopslag..." IsLoading="_isLoadingData" Alert="MessageAlert.AlertType.Info" />
-    }
-
-    @if (_loadFailed)
-    {
-        <MessageAlert Message="Fejl ved indlæsning af stillingsopslag..." Alert="MessageAlert.AlertType.Error" />
-    }
-
-    @if (_jobAdverts is null || _jobAdverts.Count() == 0)
-    {
-        if (CategoryId == 0 && SpecializationId == 0)
-        {
-            <MessageAlert Alert="MessageAlert.AlertType.Info" FullWidth="false" Message="Fandt ikke nogen ikke kategoriseret stillingsopslag!" />
-        }
-        else
-        {
-            if (SpecializationId == 0)
-            {
-                <MessageAlert Alert="MessageAlert.AlertType.Info" FullWidth="false" Message="Fandt ikke nogen stillingsopslag i denne uddannelsesgren!" />
-            }
-            else
-            {
-                <MessageAlert Alert="MessageAlert.AlertType.Info" FullWidth="false" Message="Fandt ikke nogen stillingsopslag fra dette speciale, i denne uddannelsesgren!" />
-            }
-        }
-    }
-    else
-    {
-        <div id="job-wrapper">
-            <div class="row row-cols-1 row-cols-sm-1 row-cols-md-2 row-cols-lg-2 row-cols-xl-3 g-4 px-3 py-3">
-                @foreach (var jobAdvert in _jobAdverts)
-                {
-                    <div class="col">
-                        <JobCardDetails @key="jobAdvert" JobAdvert="jobAdvert" />
-                    </div>
-                }
-            </div>
-        </div>
-    }
-</div>
