--- JobAgentSolutions/BlazorWebsite/Shared/Components/Details/CategoryDetails.razor
@@ -1,22 +1,0 @@
-﻿@using JobAgentClassLibrary.Common.Categories
-@using JobAgentClassLibrary.Common.Categories.Entities
-
-@if (_categoryDetails is not null)
-{
-    @_categoryDetails.Name
-}
-
-@code {
-    [Parameter]
-    public int Id { get; set; }
-
-    [Inject]
-    protected ICategoryService CategoryService { get; set; }
-
-    private ICategory _categoryDetails;
-
-    protected override async Task OnInitializedAsync()
-    {
-        _categoryDetails = await CategoryService.GetCategoryByIdAsync(Id);
-    }
-}
