--- JobAgentSolutions/BlazorWebsite/Shared/Components/Details/SpecializationDetails.razor
@@ -1,22 +1,0 @@
-﻿@using JobAgentClassLibrary.Common.Categories
-@using JobAgentClassLibrary.Common.Categories.Entities
-
-@if (_specializationDetails is not null)
-{
-    @_specializationDetails.Name
-}
-
-@code {
-    [Parameter]
-    public int Id { get; set; }
-
-    [Inject]
-    protected ICategoryService CategoryService { get; set; }
-
-    private ISpecialization _specializationDetails;
-
-    protected override async Task OnInitializedAsync()
-    {
-        _specializationDetails = await CategoryService.GetSpecializationByIdAsync(Id);
-    }
-}
