--- JobAgentSolutions/BlazorWebsite/Shared/Components/Diverse/MessageAlert.razor
@@ -1,62 +1,0 @@
-﻿@if (!string.IsNullOrEmpty(Message))
-{
-    <div class="@(Container)">
-        @switch (Alert)
-        {
-            case AlertType.Info:
-                <div class="alert alert-primary d-flex align-items-center" role="alert">
-                    <svg class="bi flex-shrink-0 me-2" width="24" height="24" role="img" aria-label="Info:"><use xlink:href="assets/images/sprites.svg#info-fill" /></svg>
-                    <div class="d-flex align-items-center">
-                        @Message
-
-                        @if (IsLoading)
-                        {
-                            <div class="spinner-border spinner-border-sm text-warning m-2 me-auto"></div>
-                        }
-                    </div>
-                </div>
-                break;
-            case AlertType.Success:
-                <div class="alert alert-success d-flex align-items-center" role="alert">
-                    <svg class="bi flex-shrink-0 me-2" width="24" height="24" role="img" aria-label="Success:"><use xlink:href="assets/images/sprites.svg#check-circle-fill" /></svg>
-                    <div class="d-flex align-items-center">
-                        @Message
-
-                        @if (IsLoading)
-                        {
-                            <div class="spinner-border spinner-border-sm text-success m-2 me-auto"></div>
-                        }
-                    </div>
-                </div>
-                break;
-            case AlertType.Error:
-                <div class="alert alert-danger d-flex align-items-center" role="alert">
-                    <svg class="bi flex-shrink-0 me-2" width="24" height="24" role="img" aria-label="Danger:"><use xlink:href="assets/images/sprites.svg#exclamation-triangle-fill" /></svg>
-                    <div class="d-flex align-items-center">
-                        @Message
-
-                        @if (IsLoading)
-                        {
-                            <div class="spinner-border spinner-border-sm text-danger m-2 me-auto"></div>
-                        }
-                    </div>
-                </div>
-                break;
-            case AlertType.Warning:
-                <div class="alert alert-warning d-flex align-items-center" role="alert">
-                    <svg class="bi flex-shrink-0 me-2" width="24" height="24" role="img" aria-label="Warning:"><use xlink:href="assets/images/sprites.svg#exclamation-triangle-fill" /></svg>
-                    <div class="d-flex align-items-center">
-                        @Message
-
-                        @if (IsLoading)
-                        {
-                            <div class="spinner-border spinner-border-sm text-warning m-2 me-auto"></div>
-                        }
-                    </div>
-                </div>
-                break;
-            default:
-                break;
-        }
-    </div>
-}
