--- JobAgentSolutions/BlazorWebsite/Shared/Components/Diverse/PageDocumentTitle.razor
@@ -1,11 +1,0 @@
-﻿@inject IJSRuntime JSRunTime
-
-@code {
-    [Parameter] public string DocumentTitle { get; set; }
-
-    protected override Task OnAfterRenderAsync(bool firstRender)
-    {
-        JSRunTime.InvokeVoidAsync("setDocumentTitleOnAfterRender", $"{DocumentTitle} | Din Job Agent");
-        return base.OnAfterRenderAsync(firstRender);
-    }
-}
