--- JobAgentSolutions/BlazorWebsite/Shared/Components/Diverse/PageHeadingTitle.razor
@@ -1,6 +1,0 @@
-﻿<h3 class="display-4">@DisplayTitle</h3>
-<hr class="mb-4" />
-
-@code {
-    [Parameter] public string DisplayTitle { get; set; }
-}
