--- JobAgentSolutions/BlazorWebsite/Shared/Components/Menus/NavbarMenu.razor
@@ -1,72 +1,0 @@
-﻿
-<nav class="navbar navbar-expand-lg navbar-light bg-light">
-    <div class="container-fluid me-2">
-        <a class="navbar-brand" href="/">
-            <svg version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" class="logo d-inline-block" x="0px" y="0px" viewBox="0 0 60 31" style="enable-background: new 0 0 60 31;" xml:space="preserve">
-            <polygon id="logo_z" class="st0"
-                     points="18.7,0.7 0.4,0.7 0.4,8.5 7.1,8.5 0.4,15.5 0.4,30.4 18.7,30.4 18.7,22.6 5,22.6 18.7,8.1 ">  
-                    </polygon>
-            <path id="logo_b" class="st0"
-                  d="M36.1,15.5L36.1,15.5L36.1,15.5c2.5-1.4,3.3-4.1,3.3-6.5c0-4.2-3.2-8.4-9.4-8.4h-8.8v7.8h5.5h0.9H29
-                        c1.4,0,2.1,1.1,2.1,2.2s-0.7,2.2-2.1,2.2h-1.4h-0.9h-5.5v2.8v2.8h5.5h0.9H29c1.4,0,2.1,1.1,2.1,2.2s-0.7,2.2-2.1,2.2h-1.4h-0.9h-5.5
-                        v7.8H30c6.2,0,9.4-4.2,9.4-8.4C39.4,19.6,38.5,16.9,36.1,15.5z">
-    
-                  </path>
-            <path id="logo_upper_c" class="st0"
-                  d="M46.4,8.1c2.9,0,5.1,1.9,5.8,4.4h0.5L59.7,7C57,2.8,52.3,0.1,46.4,0.1c-1.6,0-3.2,0.2-4.7,0.7v9.3
-                        C42.9,8.8,44.5,8.1,46.4,8.1z">
-    
-                  </path>
-            <path id="logo_lower_c" class="st0"
-                  d="M52.8,18.4h-0.5c-0.7,2.6-2.9,4.4-5.9,4.4c-2,0-3.6-0.8-4.7-2v9.3c1.5,0.4,3.1,0.7,4.7,0.7
-                        c5.7,0,10.7-2.8,13.4-7.2L52.8,18.4z">    
-                  </path>
-                </svg>
-            <span class="align-middle">Din Job Agent</span>
-        </a>
-
-        <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#mainNavbar" aria-controls="mainNavbar" aria-expanded="false" aria-label="Toggle navigation">
-            <span class="navbar-toggler-icon"></span>
-        </button>
-
-        <div class="collapse navbar-collapse" id="mainNavbar">
-            <div class="navbar-nav ms-auto">
-                <div class="user-profile-container d-flex">
-                    <div class="dropdown">
-                        <a href="#user-profile" role="button" id="dropdownUserProfile" data-bs-toggle="dropdown" data-bs-auto-close="false" aria-expanded="false" data-bs-reference="parent">
-                            <span class="user-profile-icon menu-icons @_userProfileState" @onclick="OnClick_UserProfileIcon">
-                                <i class="far fa-user-circle"></i>
-                            </span>
-                        </a>
-
-                        <div id="user-profile-dropdown" class="dropdown-menu dropdown-menu dropdown-menu-lg-end" aria-labelledby="dropdownUserProfile">
-                            <AuthorizeView Context="authorizeContext">
-                                <Authorized>
-                                    <UserProfileComponent />
-                                </Authorized>
-                                <NotAuthorized>
-                                    <LoginComponent />
-                                </NotAuthorized>
-                            </AuthorizeView>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-</nav>
-
-@code {
-    private string _userProfileState = string.Empty;
-
-    private void OnClick_UserProfileIcon()
-    {
-        if (string.IsNullOrEmpty(_userProfileState))
-        {
-            _userProfileState = "active";
-            return;
-        }
-
-        _userProfileState = string.Empty;
-    }
-}
