--- JobAgentSolutions/BlazorWebsite/Shared/Components/Modals/CompanyModals/CreateCompanyModal.razor
@@ -1,57 +1,0 @@
-﻿@attribute [Authorize]
-@using BlazorWebsite.Shared.Components.Diverse; 
-
-<div class="modal fade" id="ModalCreateCompany" tabindex="-1" role="dialog" aria-hidden="true" data-bs-backdrop="static" data-bs-keyboard="false">
-    <div class="modal-dialog modal-dialog-centered modal-xl" role="document">
-        <div class="modal-content">
-
-            <EditForm EditContext="_editContext" OnValidSubmit="OnValidSubmit_CreateCompany">
-
-                <div class="modal-header">
-                    <h5 class="modal-title">Opret ny Virksomhed</h5>
-                </div>
-
-                @if (_showError)
-                {
-                    <MessageAlert Alert="MessageAlert.AlertType.Error" Message="@_errorMessage" />
-                }
-
-                <div class="modal-body">
-                    <ValidationSummary />
-
-                    <div class="row">
-                        <div class="col">
-                            <label for="regInputCompanyIdentity" class="admin-label col-form-label col-form-label-sm">Virksomhedens CVR : </label>
-                            <input id="regInputCompanyIdentity" type="number" class="form-control form-control-sm" placeholder="Indtast CVR nr." @bind="_companyModel.CVR" />
-                        </div>
-                    </div>
-
-                    <hr />
-
-                    <div class="row">
-                        <div class="col">
-                            <label for="regInputName" class="admin-label col-form-label col-form-label-sm">Virksomhedens Navn :</label>
-                            <input id="regInputName" type="text" class="form-control form-control-sm" placeholder="Indtast virksomhedens navn.." @bind="_companyModel.Name" />
-                        </div>
-                    </div>
-
-                    <hr />
-
-                    <div class="row">
-                        <div class="col">
-                            <label for="regInputCompanyUrl" class="admin-label col-form-label col-form-label-sm">Navn på Virksomhedens Kontaktperson :</label>
-                            <input id="regInputCompanyUrl" type="text" class="form-control form-control-sm" placeholder="Indtast link til virksomhedens hjemmeside.." @bind="_companyModel.ContactPerson" />
-                        </div>
-                    </div>
-                </div>
-
-                <div class="modal-footer">
-                    <button type="button" class="btn btn-danger" data-bs-dismiss="modal" @onclick="CancelRequest">Annuller</button>
-                    <button type="submit" class="btn btn-info" disabled="@_isProcessing">Opret Virksomhed</button>
-                </div>
-
-            </EditForm>
-
-        </div>
-    </div>
-</div>
