--- JobAgentSolutions/BlazorWebsite/Shared/Components/Modals/CompanyModals/EditCompanyModal.razor
@@ -1,62 +1,0 @@
-﻿@using BlazorWebsite.Shared.Components.Diverse;
- 
-<div class="modal fade" id="ModalEditCompany" tabindex="-1" role="dialog" aria-labelledby="EditCompanyModalTitel" aria-hidden="true" data-bs-backdrop="static" data-bs-keyboard="false">
-    <div class="modal-dialog modal-dialog-centered modal-xl" role="document">
-        <div class="modal-content">
-
-            @if (_companyModel != null)
-            {
-
-                <EditForm Model="_companyModel" OnValidSubmit="OnValidSubmit_UpdateCompany">
-
-                    <div class="modal-header">
-                        <h5 class="modal-title" id="EditCompanyModalTitel">Virksommhedens Information :</h5>
-                    </div>
-
-                    @if (_showError)
-                    {
-                        <MessageAlert Alert="MessageAlert.AlertType.Error" Message="@_errorMessage" />
-                    }
-
-                    <div class="modal-body">
-
-                        <DataAnnotationsValidator />
-                        <ValidationSummary />
-
-                        <div class="row">
-                            <div class="col">
-                                <label for="inputCompanyIdentity" class="admin-label col-form-label col-form-label-sm">Virksomhedens CVR : <span class="font-italic small" style="color: lightcoral;">*Ændring af CVR kan muligvis skabe problemer.</span></label>
-                                <input id="inputCompanyIdentity" type="text" class="form-control form-control-sm" placeholder="Indtast CVR nr.." @bind-value="_companyModel.CVR" />
-                            </div>
-                        </div>
-
-                        <hr />
-
-                        <div class="row">
-                            <div class="col">
-                                <label for="inputName" class="admin-label col-form-label col-form-label-sm">Virksomhedens Navn :</label>
-                                <input id="inputName" type="text" class="form-control form-control-sm" placeholder="Indtast virksomhedens navn.." @bind="_companyModel.Name" />
-                            </div>
-                        </div>
-
-                        <hr />
-
-                        <div class="row">
-                            <div class="col">
-                                <label for="InputContactPerson" class="admin-label col-form-label col-form-label-sm">Kontakt Person :</label>
-                                <input id="InputContactPerson" type="text" class="form-control form-control-sm" placeholder="Indtast Kontakperson fra virksomheden.." @bind-value="_companyModel.ContactPerson" />
-                            </div>
-                        </div>
-
-                    </div>
-
-                    <div class="modal-footer">
-                        <button type="button" class="btn btn-danger" data-bs-dismiss="modal">Annuller</button>
-                        <button class="btn btn-info text-light cursor-pointer" type="submit" disabled="@_isProcessing">Gem ændringer</button>
-                    </div>
-                </EditForm>
-            }
-
-        </div>
-    </div>
-</div>
