--- JobAgentSolutions/BlazorWebsite/Shared/Components/Modals/JobAdvertModals/CreateJobAdvertModal.razor
@@ -1,130 +1,0 @@
-﻿@attribute [Authorize(Roles = "Konsulent, System Administrator, Udvikler")]
-@using BlazorWebsite.Shared.Components.Diverse; 
-
-<div id="ModalCreateJobAdvert" class="modal fade" tabindex="-1" role="dialog" aria-hidden="true" data-bs-backdrop="static" data-bs-keyboard="false">
-    <div class="modal-dialog modal-xl" role="document">
-        <div class="modal-content">
-
-            <MessageAlert Message="@(_errorMessage)" Alert="MessageAlert.AlertType.Error" />
-
-            @if (_isLoading)
-            {
-                <MessageAlert IsLoading="@_isLoading" Alert="MessageAlert.AlertType.Info" Message="Vent venligst, henter nødvendig information.." />
-            }
-            else
-            {
-                <EditForm EditContext="_editContext" OnValidSubmit="OnValidSubmit_CreateJobAdvertAsync">
-
-                    <div class="modal-header">
-                        <h5 class="modal-title">Nyt Stillingsopslag :</h5>
-                    </div>
-
-                    <div class="modal-body">
-
-                        <div class="row">
-                            <div class="col">
-                                <label for="regInputTitle" class="admin-label">Stillingsopslagets Titel :</label>
-                                <input id="regInputTitle" class="form-control form-control-sm" type="text" @bind="_jobAdvertModel.Title" placeholder="Eksempel : Kager på en ny måde .." />
-                                <ValidationMessage For="@(() => _jobAdvertModel.Title)" />
-                            </div>
-                        </div>
-
-                        <hr />
-
-                        <div class="row">
-                            <div class="col">
-                                <label for="regInputDesc" class="admin-label col-form-label col-form-label-sm">Stillingsopslags Kort-Beskrivelse :</label>
-                                <InputTextArea id="regInputDesc" class="form-control form-control-sm" style="max-height: 450px; min-height: 150px;" @bind-Value="_jobAdvertModel.Summary" placeholder="Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua."></InputTextArea>
-                                <ValidationMessage For="@(() => _jobAdvertModel.Summary)" />
-                            </div>
-                        </div>
-
-                        <hr />
-
-                        <div class="row">
-                            <div class="col">
-                                <label for="regInputRegistrationDate" class="admin-label col-form-label col-form-label-sm">Kontrakt Registreret Dato :</label>
-                                <InputDate TValue="DateTime" @bind-Value="@(_jobAdvertModel.RegistrationDateTime)" id="regInputRegistrationDate" class="form-control form-control-sm" />
-                                <ValidationMessage For="@(() => _jobAdvertModel.RegistrationDateTime)" />
-                            </div>
-                        </div>
-
-                        <hr />
-
-                        <div class="row">
-                            <div class="col">
-                                <label for="regInputCompany" class="admin-label col-form-label col-form-label-sm">Stillingsopslag Tilhører :</label>
-                                @if (_vacantJobs == null || !_vacantJobs.Any())
-                                {
-                                    <label class="col-form-label">Kunne ikke hente job sider eller er igang med at indlæse data..</label>
-                                }
-                                else
-                                {
-                                    <select id="regInputCompany" class="form-control form-control-sm" @bind="_jobAdvertModel.Id">
-                                        <option value="0" selected>---Vælg en job side fra listen som tilhører stillingsopslaget---</option>
-                                        @foreach (var vacantJob in _vacantJobs)
-                                        {
-                                            @*<option value="@vacantJob.Id">@vacantJob.Company.Name - @vacantJob.URL</option>*@
-                                        }
-                                    </select>
-                                }
-                                <ValidationMessage For="@(() => _jobAdvertModel.Id)" />
-                            </div>
-                        </div>
-
-                        <hr />
-
-                        <div class="row">
-                            <div class="col">
-                                <label for="regInputCategory" class="admin-label col-form-label col-form-label-sm">Kategori :</label>
-                                @if (_categories == null || !_categories.Any())
-                                {
-                                    <label class="form-label">Kunne ikke hente kategorier eller er igang med at indlæse data..</label>
-                                }
-                                else
-                                {
-                                    <select id="regInputCategory" class="form-control form-control-sm" @onchange="OnChange_GetSpecializationByCategoryId">
-                                        @foreach (var category in _categories)
-                                        {
-                                            if (_categories.FirstOrDefault().Id == category.Id)
-                                            {
-                                                <option value="@category.Id" selected>@category.Name</option>
-                                            }
-                                            else
-                                            {
-                                                <option value="@category.Id">@category.Name</option>
-                                            }
-                                        }
-                                    </select>
-                                }
-                                <ValidationMessage For="@(() => _jobAdvertModel.CategoryId)" />
-                            </div>
-                            @if (_sortedSpecializations != null && _sortedSpecializations.Count() != 0)
-                            {
-                                <div class="col">
-                                    <label for="regInputSpecialization" class="admin-label col-form-label col-form-label-sm">Kategori Speciale :</label>
-                                    <select id="regInputSpecialization" class="form-control form-control-sm" @bind="_jobAdvertModel.SpecializationId">
-                                        <option value="0" selected>---Vælg et speciale fra listen---</option>
-                                        @foreach (var spec in _sortedSpecializations)
-                                        {
-                                            <option value="@spec.Id">@(spec.Name)</option>
-                                        }
-                                    </select>
-                                    <ValidationMessage For="@(() => _jobAdvertModel.SpecializationId)" />
-                                </div>
-                            }
-                        </div>
-
-                    </div>
-
-                    <div class="modal-footer">
-                        <button type="button" class="btn btn-danger" data-bs-dismiss="modal" @onclick="OnClick_CancelRequest">Annuller</button>
-                        <button class="btn btn-info text-light cursor-pointer" type="submit" disabled="@_isProcessing">Opret Stillingsopslag</button>
-                    </div>
-
-                </EditForm>
-            }
-
-        </div>
-    </div>
-</div>
