--- JobAgentSolutions/BlazorWebsite/Shared/Components/Modals/JobAdvertModals/EditJobAdvertModal.razor
@@ -1,134 +1,0 @@
-﻿@attribute [Authorize(Roles = "Konsulent, System Administrator, Udvikler")]
-@using BlazorWebsite.Shared.Components.Diverse; 
-
-<div class="modal fade" id="ModalEditJobAdvert" tabindex="-1" role="dialog" aria-hidden="true" data-bs-backdrop="static" data-bs-keyboard="false">
-    <div class="modal-dialog modal-dialog-centered modal-xl" role="document">
-        <div class="modal-content">
-
-            @if (_isLoading)
-            {
-                <MessageAlert IsLoading="_isLoading" Message="Indlæser data. Vent venligst.." Alert="MessageAlert.AlertType.Info" />
-
-            }
-            <MessageAlert Message="@(_errorMessage)" Alert="MessageAlert.AlertType.Error" />
-
-            @if (Model == null)
-            {
-                <MessageAlert Alert="MessageAlert.AlertType.Warning" FullWidth="true" Message="Vent venligst, henter detailjer på stillingsopslag!" />
-            }
-            else
-            {
-                <EditForm Model="Model" OnValidSubmit="OnValidSubmit_EditJobVacancy">
-
-                    <div class="modal-header">
-
-                    </div>
-
-                    <div class="modal-body">
-                        <div class="row">
-                            <div class="col">
-                                <label for="inputTitle" class="admin-label col-form-label col-form-label-sm">Titel :</label>
-                                <input id="inputTitle" type="text" class="form-control form-control-sm" placeholder="Indtast titel" @bind-value="Model.Title" />
-                                <ValidationMessage For="@(() => Model.Title)" />
-                            </div>
-                        </div>
-                        <hr />
-                        <div class="row">
-                            <div class="col">
-                                <label for="inputDesc" class="admin-label col-form-label col-form-label-sm">Kort-Beskrivelse :</label>
-                                <textarea id="inputDesc" class="form-control form-control-sm" style="max-height: 450px; min-height: 150px;" @bind="Model.Summary"></textarea>
-                                <ValidationMessage For="@(() => Model.Summary)" />
-                            </div>
-                        </div>
-
-                        <hr />
-
-                        <div class="row">
-                            <div class="col">
-                                <label for="inputCompany" class="admin-label col-form-label col-form-label-sm">Virksomhed :</label>
-                                <select class="form-control form-control-sm" @bind="Model.Id">
-
-                                    @foreach (var vacantJob in _vacantJobs)
-                                    {
-                                        @if (vacantJob.Id == Model.Id)
-                                        {
-                                            @*<option value="@vacantJob.Id" selected>@vacantJob.Company.Name</option>*@
-                                        }
-                                        else
-                                        {
-                                            @*<option value="@vacantJob.Id">@vacantJob.Company.Name</option>*@
-                                        }
-                                    }
-                                </select>
-                                <ValidationMessage For="@(() => Model.Id)" />
-                            </div>
-                        </div>
-
-                        <hr />
-
-                        <div class="row">
-                            <div class="col">
-                                <label for="inputRegisteredDate" class="admin-label col-form-label col-form-label-sm">Opslag Registeret :</label>
-                                <InputDate id="inputRegisteredDate" @bind-Value="@Model.RegistrationDateTime" TValue="DateTime" class="form-control form-control-sm"></InputDate>
-                                <ValidationMessage For="@(() => Model.RegistrationDateTime)" />
-                            </div>
-                        </div>
-
-                        <hr />
-
-                        <div class="row">
-                            <div class="col">
-                                <label for="inputConsultantArea" class="admin-label col-form-label col-form-label-sm">Kategori :</label>
-                                <select class="form-control form-control-sm" @onchange="OnChange_SortSpecializationListByCategoryId">
-                                    @if (Model.CategoryId == 0)
-                                    {
-                                        <option value="0">Ikke Kategoriseret</option>
-                                    }
-
-                                    @foreach (var cat in _categories)
-                                    {
-                                        @if (cat.Id == Model.CategoryId)
-                                        {
-                                            <option value="@cat.Id" selected>@cat.Name</option>
-                                        }
-                                        else
-                                        {
-                                            <option value="@cat.Id">@cat.Name</option>
-                                        }
-                                    }
-                                </select>
-                                <ValidationMessage For="@(() => Model.CategoryId)" />
-                            </div>
-                            @if (_sortedSpecializations.Count != 0)
-                            {
-                                <div class="col">
-                                    <label for="inputLocation" class="admin-label col-form-label col-form-label-sm">Kategori Speciale :</label>
-                                    <select class="form-control form-control-sm" @bind="Model.SpecializationId">
-                                        @foreach (var spec in _sortedSpecializations)
-                                        {
-                                            @if (spec.Id == Model.SpecializationId)
-                                            {
-                                                <option value="@spec.Id">@spec.Name</option>
-                                            }
-                                            else
-                                            {
-                                                <option value="@spec.Id">@spec.Name</option>
-                                            }
-                                        }
-                                    </select>
-                                    <ValidationMessage For="@(() => Model.SpecializationId)" />
-                                </div>
-                            }
-                        </div>
-                    </div>
-
-                    <div class="modal-footer">
-                        <button type="button" class="btn btn-danger" data-bs-dismiss="modal" @onclick="OnClick_CancelRequest" @onclick:preventDefault>Annuller</button>
-                        <button class="btn btn-info text-light cursor-pointer" type="submit" disabled="@(_isProcessing)">Gem ændringer</button>
-                    </div>
-                </EditForm>
-            }
-
-        </div>
-    </div>
-</div>
