+++ JobAgentSolutions/BlazorWebsite/_Imports.razor
@@ -9,5 +9,3 @@
 @using Microsoft.JSInterop
 @using BlazorWebsite
 @using BlazorWebsite.Shared
-@using BlazorWebsite.Shared.Components.Details
-@using BlazorWebsite.Shared.Components.Diverse
